+++ src/input_map.rs
@@ -63,6 +63,10 @@ use std::marker::PhantomData;
 /// .insert_chord([InputKind::Keyboard(KeyCode::H),
 ///                InputKind::GamepadButton(GamepadButtonType::South),
 ///                InputKind::Mouse(MouseButton::Middle)],
+///            Action::Run)
+/// .insert_chord([InputKind::Keyboard(KeyCode::H),
+///                InputKind::GamepadButton(GamepadButtonType::South),
+///                InputKind::Mouse(MouseButton::Middle)],
 ///            Action::Hide);
 ///
 /// // Removal